+++ internal/pkgmgr/apt.go
@@ -1,0 +1,5 @@
+package pkgmgr
+
+func newAptManager(lookPath lookPathFunc) Manager {
+	return newCommandManager("apt", lookPath)
+}