+++ internal/pkgmgr/brew.go
@@ -1,0 +1,5 @@
+package pkgmgr
+
+func newBrewManager(lookPath lookPathFunc) Manager {
+	return newCommandManager("brew", lookPath)
+}