+++ internal/pkgmgr/cargo.go
@@ -1,0 +1,5 @@
+package pkgmgr
+
+func newCargoManager(lookPath lookPathFunc) Manager {
+	return newCommandManager("cargo", lookPath)
+}