+++ internal/pkgmgr/dnf.go
@@ -1,0 +1,5 @@
+package pkgmgr
+
+func newDNFManager(lookPath lookPathFunc) Manager {
+	return newCommandManager("dnf", lookPath)
+}