+++ internal/pkgmgr/go.go
@@ -1,0 +1,5 @@
+package pkgmgr
+
+func newGoManager(lookPath lookPathFunc) Manager {
+	return newCommandManager("go", lookPath)
+}