+++ internal/pkgmgr/pacman.go
@@ -1,0 +1,5 @@
+package pkgmgr
+
+func newPacmanManager(lookPath lookPathFunc) Manager {
+	return newCommandManager("pacman", lookPath)
+}