+++ internal/pkgmgr/pipx.go
@@ -1,0 +1,5 @@
+package pkgmgr
+
+func newPipxManager(lookPath lookPathFunc) Manager {
+	return newCommandManager("pipx", lookPath)
+}