+++ internal/pkgmgr/snap.go
@@ -1,0 +1,5 @@
+package pkgmgr
+
+func newSnapManager(lookPath lookPathFunc) Manager {
+	return newCommandManager("snap", lookPath)
+}